platform/bridge/pyproject.toml
[project]
name = "panel-bridge"
version = "0.1.0"
description = "WebSocket ↔ UART bridge for Thread control panels."
requires-python = ">=3.11"
dependencies = [
    "pyserial-asyncio>=0.6",
    "websockets>=11.0",
]

[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["panel_bridge*"]
